+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ```clj
 [prismatic/plumbing "0.5.3"] is available but we use "0.5.2"
+[cheshire "5.6.1"] is available but we use "5.5.0"
 ```
 
 ## 1.0.2 (27.3.2016)