+++ CHANGELOG.md
@@ -2,10 +2,24 @@
 
 * Strip nils from `:middleware`, fixes [#228](https://github.com/metosin/compojure-api/issues/228)
 * Lazily require `ring.swagger.validator` namespace in `compojure.api.swagger/validate` to allow compojure-api apps in [Google App Engine](https://cloud.google.com/appengine), Fixes [#227](https://github.com/metosin/compojure-api/issues/227). **NOTE** exluding `metosin/scjsv` will cause the validate to fail at runtime.
-* Resources now implement `compojure.response/Renderable`, so that they can be used with endpoint macros like `ANY`
 * **BREAKING**: If a resource doesn't define a handler for a given `request-method` or for top-level, nil is returned (instead of throwing exeption)
+* Resource-routing is done with `context`. Trying to return a `compojure.api.routing/Route` from an endpoint like `ANY` will throw descriptive (runtime-)exception.
 
+```clj
+(context "/hello" []
+  (resource
+    {:description "hello-resource"
+     :responses {200 {:schema {:message s/Str}}}
+     :post {:summary "post-hello"
+            :parameters {:body-params {:name s/Str}}
+            :handler (fnk [[:body-params name]]
+                       (ok {:message (format "hello, %s!" name)}))}
+     :get {:summary "get-hello"
+           :parameters {:query-params {:name s/Str}}
+           :handler (fnk [[:query-params name]]
+                      (ok {:message (format "hello, %s!" name)}))}}))
 ```
+
 [metosin/compojure-api "1.0.3" :exclusions [[metosin/scjsv]]]
 ```